+++ Packages/jp.ac.keio.sfc.sdp/Runtime/Scripts/Struct/GcKeyState.cs
@@ -9,8 +9,8 @@ namespace GameCanvas
         public bool Up { get; }
         public bool Cancelled { get; }
         /// <summary>押してからの秒数。終了したフレームまで読めます。</summary>
-        public float Duration { get; }
-        internal GcKeyState(bool down, bool held, bool up, bool cancelled, float duration)
+        public double Duration { get; }
+        internal GcKeyState(bool down, bool held, bool up, bool cancelled, double duration)
         { Down = down; Held = held; Up = up; Cancelled = cancelled; Duration = duration; }
     }
 }
